+++ src/app/bridge/steps/step_two.tsx
@@ -6,13 +6,13 @@ import { useBlockNumber, useReadContract, useWaitForTransactionReceipt } from "w
 import { WindToy } from "react-svg-spinners";
 import { useEffect, useState } from "react";
 import { ethers } from "ethers";
-import { L1BlockABI } from "@/app/bridge/util/abis";
 import { getStepThreeURL, getStepTwoURL } from "./urls";
 import { useQuery } from "@tanstack/react-query";
-import { getBlockchainState, getCoinRecordByName, getPuzzleAndSolution } from "../util/rpc";
-import { hexToString } from "../util/sig";
 import * as GreenWeb from 'greenwebjs';
 import { ConditionOpcode } from "greenwebjs/util/sexp/condition_opcodes";
+import { getBlockchainState, getCoinRecordByName, getPuzzleAndSolution } from "../drivers/rpc";
+import { getMessageSentFromXCHStepThreeData } from "../drivers/portal";
+import { L1BlockABI } from "../drivers/abis";
 
 export default function StepTwo({
   sourceChain,
@@ -45,53 +45,6 @@ export default function StepTwo({
   );
 }
 
-async function fetchXCHMessageDetailsAndNavigate(
-  router: any,
-  sourceChain: Network,
-  nonce: string,
-) {
-  const messageCoinRecord = await getCoinRecordByName(sourceChain.rpcUrl, nonce);
-  const messageCoinParentSpend = await getPuzzleAndSolution(
-    sourceChain.rpcUrl,
-    messageCoinRecord.coin.parent_coin_info,
-    messageCoinRecord.confirmed_block_index
-  );
-
-  const [_, conditionsDict, __] = GreenWeb.util.sexp.conditionsDictForSolution(
-    GreenWeb.util.sexp.fromHex(messageCoinParentSpend.puzzle_reveal.slice(2)),
-    GreenWeb.util.sexp.fromHex(messageCoinParentSpend.solution.slice(2)),
-    GreenWeb.util.sexp.MAX_BLOCK_COST_CLVM
-  )
-  var createCoinConds = conditionsDict?.get("33" as ConditionOpcode) ?? [];
-
-  createCoinConds.forEach((cond) => {
-    if(cond.vars[0] === messageCoinRecord.coin.puzzle_hash.slice(2) &&
-       cond.vars[1] === GreenWeb.util.coin.amountToBytes(messageCoinRecord.coin.amount)) {
-        const memos = GreenWeb.util.sexp.fromHex(cond.vars[2]);
-        
-        const destination_chain_id = GreenWeb.util.sexp.toHex(memos.first()).slice(2);
-        const destination = GreenWeb.util.sexp.toHex(memos.rest().first()).slice(2);
-        
-        const contents = GreenWeb.util.sexp.asAtomList(memos.rest().rest()).map((val) => {
-          if(val.length === 64) {
-            return val;
-          }
-
-          return "0".repeat(64 - val.length) + val;
-        });
-
-        router.push(getStepThreeURL({
-          sourceNetworkId: sourceChain.id,
-          destinationNetworkId: hexToString(destination_chain_id),
-          nonce,
-          source: messageCoinParentSpend.coin.puzzle_hash.slice(2),
-          destination: ethers.getAddress("0x" + destination),
-          contents
-        }));
-      }
-  })
-}
-
 
 function XCHValidationElement({
   txHash,
@@ -117,7 +70,9 @@ function XCHValidationElement({
       onConfirmation={async () => {
         const nonce = txHash;
         
-        await fetchXCHMessageDetailsAndNavigate(router, sourceChain, nonce);
+        router.push(getStepThreeURL(
+          await getMessageSentFromXCHStepThreeData(sourceChain, nonce)
+        ));
       }}
     />
   );
@@ -132,7 +87,7 @@ function XCHMempoolFollower({
   coinId: string,
   setBlock: (block: number) => void,
 }) {
-  const { data, refetch } = useQuery({
+  const { data } = useQuery({
     queryKey: ['StepTwo_fetchCoinRecordByName', coinId],
     queryFn: () => getCoinRecordByName(sourceChainRpcUrl, coinId).then((record) => record ?? false),
     enabled: true,
